pyproject.toml
[project]
name = "doc-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "phidata>=2.7",
    "openai>=1.0",
    "python-dotenv",
    "rich",
]

[project.scripts]
plume-doc = "plume_doc_agent.cli:main"

[tool.setuptools.packages.find]
include = ["plume_doc_agent*"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
